standalone_bot/requirements.txt
python-telegram-bot==21.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
